feat: chat viewer and text messages
Created chats and scrollable chats previews on the left panel, they can only be opened by clicking on the chat image (will be fixed).
Added scrollable messages (text only) and message sending simulation.
Moved constants in a separate file

There is a bug when the window is stretched vertically: the scrolling is inconsistent